refactor: 불러온 이미지 파일명 변경 및 위치 이동
1. 엄밀히 말하면 image'loader'가 아니기 때문에 이름 변경
2. 같은 이유로 util에 있기에는 무리가 있다고 생각
