--- src/util/ImageLoader.js
@@ -1,18 +1,0 @@
-import asianImg from "../../templates/category-asian.png";
-import chineseImg from "../../templates/category-chinese.png";
-import etcImg from "../../templates/category-etc.png";
-import japaneseImg from "../../templates/category-japanese.png";
-import koreanImg from "../../templates/category-korean.png";
-import westernImg from "../../templates/category-western.png";
-import addButton from "../../templates/add-button.png";
-
-export const IMAGE = {
-  아시안: asianImg,
-  양식: westernImg,
-  일식: japaneseImg,
-  한식: koreanImg,
-  중식: chineseImg,
-  기타: etcImg,
-  전체: etcImg,
-  ADD_BTN: addButton,
-};
